refactor(Cars): extract getting winner feature from RacingGame
