+++ src/main/java/racingcar/domain/Cars.java
@@ -4,10 +4,11 @@ import racingcar.util.RandomUtils;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class Cars {
 
-    private static final List<Car> cars = new ArrayList<>();
+    private final List<Car> cars = new ArrayList<>();
 
     public List<Car> findAllCars() {
         return cars;
@@ -25,5 +26,19 @@ public class Cars {
         for (Car car : cars) {
             car.goOrNot(RandomUtils.generateNumber());
         }
+    }
+
+    public List<Car> getWinners() {
+        int maxPosition = getMaxPosition();
+
+        return cars.stream()
+                .filter((car) -> car.getPosition() == maxPosition)
+                .collect(Collectors.toList());
+    }
+
+    private int getMaxPosition() {
+        return cars.stream()
+                .mapToInt(Car::getPosition)
+                .max().orElseThrow(null);
     }
 }
